// packages/reldens-schemas/src/generated/constants.ts
export const RELDENS_CONSTANTS = {
    "ActionsConst": {
        "BATTLE_TYPE_PER_TARGET": "bt",
        "BATTLE_TYPE_GENERAL": "bg",
        "BATTLE_ENDED": "bend",
        "TARGET_POSITION": "tgp",
        "TARGET_PLAYER": "tga",
        "TARGET_OBJECT": "tgo",
        "FULL_SKILLS_LIST": "fkl",
        "ACTION": "action",
        "DATA_OBJECT_KEY_TARGET": "t",
        "DATA_OBJECT_KEY_OWNER": "o",
        "DATA_TARGET_TYPE": "tT",
        "DATA_TARGET_KEY": "tK",
        "DATA_OWNER_TYPE": "oT",
        "DATA_OWNER_KEY": "oK",
        "DATA_TYPE_VALUE_ENEMY": "e",
        "DATA_TYPE_VALUE_PLAYER": "p",
        "DATA_TYPE_VALUE_OBJECT": "o",
        "EXTRA_DATA.KEY": "sked",
        "EXTRA_DATA.SKILL_DELAY": "sd",
        "DEFAULT_HIT_ANIMATION_KEY": "default_hit",
        "ACTIONS.SUFFIX.ATTACK": "_atk",
        "ACTIONS.SUFFIX.EFFECT": "_eff",
        "ACTIONS.SUFFIX.HIT": "_hit",
        "MESSAGE.DATA.LEVEL": "lvl",
        "MESSAGE.DATA.EXPERIENCE": "exp",
        "MESSAGE.DATA.CLASS_PATH_LABEL": "lab",
        "MESSAGE.DATA.NEXT_LEVEL_EXPERIENCE": "ne",
        "MESSAGE.DATA.SKILL_LEVEL": "skl",
        "MESSAGE.DATA.LAST_ATTACK_KEY": "k",
        "MESSAGE.DATA_VALUES.NAMESPACE": "actions",
        "MESSAGE.DATA_VALUES.lvl": "level",
        "MESSAGE.DATA_VALUES.exp": "experience",
        "MESSAGE.DATA_VALUES.lab": "classPathLabel",
        "MESSAGE.DATA_VALUES.ne": "nextLevelExperience",
        "MESSAGE.DATA_VALUES.skl": "skillLevel",
        "SELECTORS.LEVEL_LABEL": ".level-container .level-label",
        "SELECTORS.CURRENT_EXPERIENCE": ".experience-container .current-experience",
        "SELECTORS.NEXT_LEVEL_EXPERIENCE": ".experience-container .next-level-experience",
        "SELECTORS.PLAYER_CREATE_FORM": "#player-create-form",
        "SELECTORS.UI_PLAYER_EXTRAS": "#ui-player-extras",
        "SELECTORS.PLAYER_CREATION_ADDITIONAL_INFO": ".player-creation-additional-info",
        "SELECTORS.PLAYER_SELECTION_ADDITIONAL_INFO": ".player-selection-additional-info",
        "SELECTORS.CLASS_PATH_LABEL": ".class-path-container .class-path-label",
        "SELECTORS.SKILLS_CONTAINER": ".skills-container",
        "SNIPPETS.PREFIX": "actions.",
        "SNIPPETS.SELECT_CLASS_PATH": "actions.selectClassPath",
        "SNIPPETS.EXPERIENCE_LABEL": "actions.experienceLabel",
        "SNIPPETS.LEVEL": "actions.currentLevel",
        "SNIPPETS.CLASS_PATH_LABEL": "actions.classPathLabel",
        "SNIPPETS.NEXT_LEVEL_EXPERIENCE": "actions.nextLevelExperience",
        "SNIPPETS.EXPERIENCE": "actions.experience"
    },
// ...
    "GameConst": {
        "START_GAME": "s",
        "ACTION_KEY": "act",
        "CREATE_PLAYER": "cp",
        "CREATE_PLAYER_RESULT": "cps",
        "CHANGING_SCENE": "cgs",
        "CHANGED_SCENE": "cs",
        "RECONNECT": "r",
        "ROOM_GAME": "room_game",
        "ROOM_NAME_MAP": "Map",
        "SCENE_PRELOADER": "ScenePreloader",
        "PLAYER_STATS": "ps",
        "ICON_STATS": "player-stats",
        "CLIENT_JOINED": "cj",
        "UI": "ui",
        "CLOSE_UI_ACTION": "closeUi",
        "TYPE_PLAYER": "pj",
        "GAME_OVER": "go",
        "REVIVED": "rv",
        "BUTTON_OPTION": "btn-opt",
        "UI_BOX": "box",
        "UI_CLOSE": "-close",
        "UI_OPEN": "-open",
        "UP": "up",
        "LEFT": "left",
        "DOWN": "down",
        "RIGHT": "right",
        "STOP": "stop",
        "POINTER": "mp",
        "ARROW_DOWN": "ard",
        "IMAGE_PLAYER": "player",
        "IMAGE_PLAYER_BASE": "player-base.png",
        "ACTIONS.LOGIN_UPDATE_ERROR": "luer",
        "STATUS.ACTIVE": 1,
        "STATUS.DISABLED": 2,
        "STATUS.DEATH": 3,
        "STATUS.AVOID_INTERPOLATION": 4,
        "STRUCTURE.DEFAULT": "default",
        "STRUCTURE.ASSETS": "assets",
        "STRUCTURE.CSS": "css",
        "STRUCTURE.DIST": "dist",
        "STRUCTURE.THEME": "theme",
        "STRUCTURE.LIB": "lib",
        "STRUCTURE.SERVER": "server",
        "STRUCTURE.CLIENT": "client",
        "STRUCTURE.PLUGINS": "plugins",
        "STRUCTURE.INDEX": "index.html",
        "STRUCTURE.SCSS_FILE": "styles.scss",
        "STRUCTURE.CSS_FILE": "styles.css",
        "STRUCTURE.ADMIN": "admin",
        "STRUCTURE.TEMPLATES": "templates",
        "STRUCTURE.ADMIN_JS_FILE": "reldens-admin-client.js",
        "STRUCTURE.ADMIN_CSS_FILE": "reldens-admin-client.css",
        "STRUCTURE.INSTALLER_FOLDER": "install",
        "STRUCTURE.INSTALL_LOCK": "install.lock",
        "ROUTE_PATHS.DISCONNECT_USER": "/reldens-disconnect-user",
        "ROUTE_PATHS.TERMS_AND_CONDITIONS": "/terms-and-conditions",
        "ROUTE_PATHS.MAILER": "/reldens-mailer-enabled",
        "ROUTE_PATHS.FIREBASE": "/reldens-firebase",
        "SELECTORS.BODY": "body",
        "SELECTORS.CANVAS": "CANVAS",
        "SELECTORS.INPUT": "input",
        "SELECTORS.FORMS_CONTAINER": ".forms-container",
        "SELECTORS.REGISTER_FORM": "#register-form",
        "SELECTORS.GUEST_FORM": "#guest-form",
        "SELECTORS.LOGIN_FORM": "#login-form",
        "SELECTORS.FORGOT_PASSWORD_FORM": "#forgot-form",
        "SELECTORS.PLAYER_CREATE_FORM": ".player-create-form",
        "SELECTORS.PLAYER_SELECTION": "#player-selection",
        "SELECTORS.PLAYER_SELECTION_FORM": "#player-selector-form",
        "SELECTORS.PLAYER_SELECT_ELEMENT": "#player-select-element",
        "SELECTORS.PLAYER_SELECTION_ADDITIONAL_INFO": ".player-selection-additional-info",
        "SELECTORS.PLAYER_STATS_CONTAINER": "#player-stats-container",
        "SELECTORS.FULL_SCREEN_BUTTON": "#full-screen-btn",
        "SELECTORS.RESPONSE_ERROR": ".response-error",
        "SELECTORS.LOADING_CONTAINER": ".loading-container",
        "SELECTORS.REGISTRATION.PASSWORD": "#reg-password",
        "SELECTORS.REGISTRATION.RE_PASSWORD": "#reg-re-password",
        "SELECTORS.REGISTRATION.EMAIL": "#reg-email",
        "SELECTORS.REGISTRATION.USERNAME": "#reg-username",
        "SELECTORS.GUEST.USERNAME": "#guest-username",
        "SELECTORS.LOGIN.USERNAME": "#username",
        "SELECTORS.LOGIN.PASSWORD": "#password",
        "SELECTORS.FORGOT_PASSWORD.EMAIL": "#forgot-email",
        "SELECTORS.FORGOT_PASSWORD.CONTAINER": ".forgot-password-container",
        "SELECTORS.TERMS.BOX": "#terms-and-conditions",
        "SELECTORS.TERMS.CONTAINER": ".terms-and-conditions-container",
        "SELECTORS.TERMS.LINK_CONTAINER": ".terms-and-conditions-link-container",
        "SELECTORS.TERMS.LINK": ".terms-and-conditions-link",
        "SELECTORS.TERMS.ACCEPT": "#accept-terms-and-conditions",
        "SELECTORS.TERMS.ACCEPT_LABEL": ".accept-terms-and-conditions-label",
        "SELECTORS.TERMS.HEADING": ".terms-heading",
        "SELECTORS.TERMS.BODY": ".terms-body",
        "SELECTORS.TERMS.CLOSE": ".terms-and-conditions-accept-close",
        "SELECTORS.GAME_CONTAINER": ".game-container",
        "SELECTORS.BUTTONS_CLOSE": ".box-close",
        "CLASSES.HIDDEN": "hidden",
        "CLASSES.GAME_STARTED": "game-started",
        "CLASSES.GAME_ERROR": "game-error",
        "CLASSES.GAME_ENGINE_STARTED": "game-engine-started",
        "CLASSES.FULL_SCREEN_ON": "full-screen-on",
        "MESSAGE.DATA_VALUES.NAMESPACE": "game",
        "LABELS.YES": "Yes",
        "LABELS.NO": "No",
        "ANIMATIONS_TYPE.SPRITESHEET": "spritesheet",
        "FILES.EXTENSIONS.PNG": ".png",
        "GRAPHICS.FRAME_WIDTH": 32,
        "GRAPHICS.FRAME_HEIGHT": 32,
        "SHOW_PLAYER_TIME.NONE": -1,
        "SHOW_PLAYER_TIME.ONLY_OWN_PLAYER": 0,
        "SHOW_PLAYER_TIME.ALL_PLAYERS": 2
    },
// ...
    "ObjectsConst": {
        "OBJECT_ANIMATION": "oa",
        "OBJECT_INTERACTION": "oi",
        "TYPE_OBJECT": "obj",
        "TYPE_ANIMATION": "anim",
        "TYPE_NPC": "npc",
        "TYPE_ENEMY": "enemy",
        "TYPE_TRADER": "trader",
        "TYPE_DROP": "drop",
        "DYNAMIC_ANIMATION": "dyn",
        "MESSAGE.DATA_VALUES.NAMESPACE": "objects",
        "EVENT_PREFIX.BASE": "bo",
        "EVENT_PREFIX.ANIMATION": "ao",
        "EVENT_PREFIX.DROP": "dep",
        "EVENT_PREFIX.ENEMY": "eo",
        "EVENT_PREFIX.NPC": "npc",
        "EVENT_PREFIX.TRADER": "tnpc",
        "SNIPPETS.PREFIX": "objects.",
        "SNIPPETS.NPC_INVALID": "objects.npcInvalid",
        "SNIPPETS.TRADER.CONTENT": "objects.trader.content",
        "SNIPPETS.TRADER.OPTIONS.BUY": "objects.trader.options.buy",
        "SNIPPETS.TRADER.OPTIONS.SELL": "objects.trader.options.sell",
        "SNIPPETS.TRADER.BUY_CONFIRMED": "objects.trader.buyConfirmed",
        "SNIPPETS.TRADER.SELL_CONFIRMED": "objects.trader.sellConfirmed",
        "DEFAULTS.BASE_OBJECT.CONTENT": "",
        "DEFAULTS.TRADER_OBJECT.INVENTORY_MAP.buy": "A",
        "DEFAULTS.TRADER_OBJECT.INVENTORY_MAP.sell": "B",
        "DEFAULTS.TRADER_OBJECT.OPTIONS.BUY": "buy",
        "DEFAULTS.TRADER_OBJECT.OPTIONS.SELL": "sell",
        "DEFAULTS.TARGETS.OBJECT": 0,
        "DEFAULTS.TARGETS.PLAYER": 1,
        "TRADE_ACTIONS_FUNCTION_NAME.ADD": "add",
        "TRADE_ACTIONS_FUNCTION_NAME.REMOVE": "remove",
        "TRADE_ACTIONS_FUNCTION_NAME.CONFIRM": "confirm",
        "TRADE_ACTIONS_FUNCTION_NAME.DISCONFIRM": "disconfirm",
        "TRADE_ACTIONS_FUNCTION_NAME.CANCEL": "cancel",
        "TRADE_ACTIONS.SUB_ACTION": "sub",
        "TRADE_ACTIONS.ADD": "ta",
        "TRADE_ACTIONS.REMOVE": "tr",
        "TRADE_ACTIONS.CONFIRM": "tc",
        "TRADE_ACTIONS.DISCONFIRM": "td",
        "DROPS.KEY": "drp",
        "DROPS.REMOVE": "drmv",
        "DROPS.PARAMS": "drpp",
        "DROPS.ASSET_KEY": "dk",
        "DROPS.PICK_UP_ACT": "rpu",
        "DROPS.ASSETS_PATH": "/assets/custom/sprites/",
        "DROPS.FILE": "df",
        "DROPS.TYPE": "dt"
    },
// ...
} as const;